% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskyBehaviors.R
\docType{data}
\name{riskyBehaviors}
\alias{riskyBehaviors}
\title{riskyBehaviors}
\format{
A tibble with 13,677 and 45 variables
}
\usage{
data(riskyBehaviors)
}
\description{
This dataset is a subset of the 2019 Youth Risk Behavior
Surveillance System (YRBSS) data, consisting of 45 variables related
to demographics such as sex, race, age, grade, and sexual orientation,
as well as substance use and reckless behaviors in adolescents.
However, it is important to note that this subset is intended for
teaching purposes only and should not be used to draw definitive
conclusions. This is because all variables related to weights,
strata, and PSU used in the original survey have been removed
from this subset.
}
\details{
Youth Risk Behavior Surveillance System 2019 - Risky Behaviors and Substance
Use Subset
}
\keyword{datasets}
